+++ SDL3Engine/Units/uCHXSDL3TypeHelpers.pas
@@ -1,168 +1,279 @@
 unit uCHXSDL3TypeHelpers;
-{< Unit with helpers for SDL3 types.
+{< Unit with helpers for SDL3 types, and some useful new types and method:
 
-  ToDo: Actually implement helpers for SDL3 Types.
+  - **CHXSDLFSegment**: Struct with a pair of `SDL_FPoint`, sometimes it's
+    usefull store it's endpoint instead using `TSDL_FRect` and calculate
+    end point every time adding `X + W` and `Y + H`.
+
+
 
   (C) 2026 Chixpy https://github.com/Chixpy
 }
-{$mode ObjFPC}{$H+}
-{$inline ON}
+{$mode ObjFPC}{$H+}{$inline ON}{$modeswitch ADVANCEDRECORDS}
 
 interface
 
 uses
-  SysUtils, CTypes, Math, FGL,
+  SysUtils, CTypes, Math,
   SDL3;
 
+const
+  kSigmaCFloat = 1e-4; // Usual Sigma value for Single
+  kInv255 = 0.00392156863; // 1 / 255
+
 type
+
+  TCHXSDLFSegment = record
+    P1, P2: TSDL_FPoint;
+
+    procedure Init(const aP1, aP2: TSDL_FPoint); overload; inline;
+    procedure Init(const X1, Y1, X2, Y2: CFloat); overload; inline;
+
+  end;
+  
   {
     Dynamic lists.
+
+    Useful for inheriting a create methods for batch operation.
+
+    For example, rotate all points in the list, calculating sen and cos only
+    one time for all.
+
+    ToDo: Try Generics.Collections as FGL don't work and actually uses an array.
   }
 (*
-  cSDLColorList = specialize TFPGList<TSDL_Color>;
-  //< List of TSDL_Color.
-  cSDLPointList = specialize TFPGList<TSDL_Point>;
-  //< List of TSDL_Point.
-  cSDLFPointList = specialize TFPGList<TSDL_FPoint>;
+  cSDLColorList = specialize TList<TSDL_FColor>;
+  //< List of TSDL_FColor.
+  cSDLFPointList = specialize TList<TSDL_FPoint>;
   //< List of TSDL_FPoint.
-  cSDLRectList = specialize TFPGList<TSDL_Rect>;
-  //< List of TSDL_Rect.
-  cSDLFRectList = specialize TFPGList<TSDL_FRect>;
+  cSDLFRectList = specialize TList<TSDL_FRect>;
   //< List of TSDL_FRect.
-  cSDLVertexList = specialize TFPGList<TSDL_Vertex>;
+  cSDLVertexList = specialize TList<TSDL_Vertex>;
   //< List of TSDL_Vertex.
-
-operator = (const a, b: TSDL_Color): Boolean;
-operator = (const a, b: TSDL_Point): Boolean;
-operator = (const a, b: TSDL_FPoint): Boolean;
-operator = (const a, b: TSDL_Rect): Boolean;
-operator = (const a, b: TSDL_FRect): Boolean;
-operator = (const a, b: TSDL_Vertex): Boolean;
 *)
 
   {
     Dynamic Arrays.
+
+    Many of the methods can actually have open array parameters, but some
+    of them modify and returns arrays.
   }
-  
-  TSDLColorDynArray = specialize TArray<TSDL_Color>;
-  //< Dynamic array of TSDL_Color.
-  TSDLPointDynArray = specialize TArray<TSDL_Point>;
-  //< Dynamic array of TSDL_Point.
-  TSDLFPointDynArray = specialize TArray<TSDL_FPoint>;
+
+  TSDLFColorDynArray = Array of TSDL_FColor;
+  //< Dynamic array of TSDL_FColor.
+  TSDLFPointDynArray = Array of TSDL_FPoint;
   //< Dynamic array of TSDL_FPoint.
-  TSDLRectDynArray = specialize TArray<TSDL_Rect>;
-  //< Dynamic array of TSDL_Rect.
-  TSDLFRectDynArray = specialize TArray<TSDL_FRect>;
+  TSDLFRectDynArray = Array of TSDL_FRect;
   //< Dynamic array of TSDL_FRect.
-  TSDLVertexDynArray = specialize TArray<TSDL_Vertex>;
+  TSDLVertexDynArray = Array of TSDL_Vertex;
   //< Dynamic array of TSDL_Vertex.
+
+{
+  Helpers for SDL types
+}
+  TSDLFColorH = record helper for TSDL_FColor
+    procedure Init(const aR, aG, aB: CFloat; const aA: CFloat = 1); overload;
+      inline;
+    procedure Init(const Grey: CFloat; const aA: CFloat = 1); overload; inline;
+
+    procedure InitByte(const aR, aG, aB: Byte; const aA: Byte = 255); overload;
+      inline;
+    procedure InitByte(const Grey: Byte; const aA: Byte = 255); overload;
+      inline;
+
+    function IsEqual(const aColor: TSDL_FColor): Boolean;
+
+    // operator = (const C1, C2: TSDL_FColor): Boolean; overload;
+
+  end;
+
+  TSDLFPointH = record helper for TSDL_FPoint
+    procedure Init(const aX, aY: CFloat); overload; inline;
+
+    // operator = (const P1, P2: TSDL_FPoint): Boolean; overload; inline;
+  end;
+
+  TSDLFRectH = record helper for TSDL_FRect
+    procedure Init(const aX, aY, aW, aH: CFloat); overload; inline;
+
+    procedure Shrink(const aSize: CFloat);
+
+    // operator = (const R1, R2: TSDL_FRect): Boolean; overload; inline;
+  end;
+
+{
+  Operator overloading.
+
+  Ideally they would be in helpers as class operators...
+}
+
+operator = (const C1, C2: TSDL_FColor): Boolean; overload;
+operator = (const P1, P2: TSDL_FPoint): Boolean; overload; inline;
+operator = (const R1, R2: TSDL_FRect): Boolean; overload; inline;
+operator = (const V1, V2: TSDL_Vertex): Boolean; overload; inline;
 
 {
   Type creation functions. Useful to use them as parameters when calling
     a function without creating a temporal variable.
 }
 
-function SDLColor(const r : CUInt8; const g : CUInt8; const b : CUInt8;
-  const a : CUInt8 = 255) : TSDL_Color; inline;
-{< Create a TSDL_Color from red, green, blue and alpha values.}
+function SDLFColor(const R, G, B: CFloat; const A : CFloat = 1): TSDL_FColor;
+  overload; inline;
+{< Create a TSDL_FColor from Red, Green, Blue and Alpha values.}
 
-function SDLPoint(const x : CInt; const y : CInt) : TSDL_Point; inline;
-{< Create a TSDL_Point.}
-function SDLFPoint(const x : CFloat; const y : CFloat) : TSDL_FPoint; inline;
+function SDLFColor(const Grey: CFloat; const A : CFloat = 1): TSDL_FColor;
+  overload; inline;
+{< Create a TSDL_FColor with a Grey value.}
+
+function SDLFPoint(const X, Y: CFloat): TSDL_FPoint; inline;
 {< Create a TSDL_FPoint.}
 
-function SDLRect(const x : CUInt; const y : CUInt; const w : CUInt;
-  const h : CUInt) : TSDL_Rect; inline;
-{< Create a TSDL_Rect.}
-function SDLFRect(const x : CFloat; const y : CFloat; const w : CFloat;
-  const h : CFloat) : TSDL_FRect; inline;
+function SDLFRect(const X, Y, W, H: CFloat): TSDL_FRect; inline;
 {< Create a TSDL_FRect.}
 
+function SDLFSegment(const aP1, aP2: TSDL_FPoint): TCHXSDLFSegment;
+  overload; inline;
+function SDLFSegment(const X1, Y1, X2, Y2: CFloat): TCHXSDLFSegment;
+  overload; inline;
 
+(*
 function SDLColor2Str(aColor : TSDL_Color) : String; inline;
 {< Write a TSDL_Color to a String.}
+
 function Str2SDLColor(aColor : String) : TSDL_Color;
 {< Write a TSDL_Color to a String.}
+*)
 
 implementation
 
-function SDLColor(const r : CUInt8; const g : CUInt8; const b : CUInt8;
-  const a : CUInt8) : TSDL_Color;
+// TCHXSDLFSegment
+
+procedure TCHXSDLFSegment.Init(const aP1, aP2: TSDL_FPoint);
 begin
-  Result.r := r;
-  Result.g := g;
-  Result.b := b;
-  Result.a := a;
+  Self.P1 := aP1; Self.P2 := aP2;
 end;
 
-function SDLPoint(const x : CInt; const y : CInt) : TSDL_Point;
+procedure TCHXSDLFSegment.Init(const X1, Y1, X2, Y2: CFloat);
 begin
-  Result.x := x;
-  Result.y := y;
+  Self.P1.X := X1; Self.P1.Y := Y1;
+  Self.P2.X := X2; Self.P2.Y := Y2;
 end;
 
-function SDLFPoint(const x : CFloat; const y : CFloat) : TSDL_FPoint;
+// TSDLFColorH
+
+procedure TSDLFColorH.Init(const aR, aG, aB, aA: CFloat);
 begin
-  Result.x := x;
-  Result.y := y;
+  Self.R := aR; Self.G := aG; Self.B := aB; Self.A := aA;
 end;
 
-function SDLRect(const x : CUInt; const y : CUInt; const w : CUInt;
-  const h : CUInt) : TSDL_Rect;
+procedure TSDLFColorH.Init(const Grey, aA: CFloat);
 begin
-  Result.x := x;
-  Result.y := y;
-  Result.w := w;
-  Result.h := h;
+  Self.R := Grey; Self.G := Grey; Self.B := Grey; Self.A := aA;
 end;
 
-function SDLFRect(const x : CFloat; const y : CFloat; const w : CFloat;
-  const h : CFloat) : TSDL_FRect;
+procedure TSDLFColorH.InitByte(const aR, aG, aB, aA: Byte);
 begin
-  Result.x := x;
-  Result.y := y;
-  Result.w := w;
-  Result.h := h;
+  Self.Init(aR * kInv255, aG * kInv255, aB * kInv255, aA * kInv255);
+end;
+procedure TSDLFColorH.InitByte(const Grey, aA: Byte);
+var
+ aGrey: CFloat;
+begin
+  aGrey := Grey * kInv255;
+  Self.Init(aGrey, aGrey, aGrey, aA * kInv255);
+end;
+
+function TSDLFColorH.IsEqual(const aColor: TSDL_FColor): Boolean;
+begin
+  // If both are totally transparent are considered the same always.
+  if IsZero(Self.A) and IsZero(aColor.A) then Exit(True);
+  Result := SameValue(Self.R, aColor.R) and SameValue(Self.G, aColor.G) 
+      and SameValue(Self.B, aColor.B) and SameValue(Self.A, aColor.A);
+end;
+
+// TSDLFPointH
+procedure TSDLFPointH.Init(const aX, aY: CFloat);
+begin
+  Self.X := aX; Self.Y := aY;
+end;
+
+// TSDLFRectH
+
+procedure TSDLFRectH.Init(const aX, aY, aW, aH: CFloat);
+begin
+  Self.X := aX; Self.Y := aY; Self.W := aW; Self.H := aH;
+end;
+
+procedure TSDLFRectH.Shrink(const aSize: CFloat);
+begin
+  X += aSize; Y += aSize; W -= (aSize + aSize); H -= (aSize + aSize);
+end;
+
+// Operators
+
+operator = (const C1, C2: TSDL_FColor): Boolean;
+begin
+  // If both are totally transparent are considered the same always.
+  if IsZero(C1.A) and IsZero(C2.A) then Exit(True);
+  Result := SameValue(C1.R, C2.R) and SameValue(C1.G, C2.G) 
+    and SameValue(C1.B, C2.B) and SameValue(C1.A, C2.A);
+end;
+
+operator = (const P1, P2: TSDL_FPoint): Boolean;
+begin
+  // ToDo: ¿Hacer estricto?
+  Result := SameValue(P1.X, P2.X) and SameValue(P1.Y, P2.Y);
+end;
+
+operator = (const R1, R2: TSDL_FRect): Boolean;
+begin
+  // ToDo: ¿Hacer estricto?
+  Result := SameValue(R1.X, R2.X) and SameValue(R1.Y, R2.Y) 
+    and SameValue(R1.W, R2.W) and SameValue(R1.H, R2.H)
+end;
+
+operator = (const V1, V2: TSDL_Vertex): Boolean;
+begin
+  Result := (V1.Position = V2.Position) and (V1.Color = V2.Color) 
+    and (V1.Tex_Coord = V2.Tex_Coord);
+end;
+
+
+
+// Type creation functions
+
+function SDLFColor(const R, G, B, A : CFloat) : TSDL_FColor;
+begin
+  Result.R := R; Result.G := G; Result.B := B; Result.A := A;
+end;
+
+function SDLFColor(const Grey, A : CFloat): TSDL_FColor;
+begin
+  Result.R := Grey; Result.G := Grey; Result.B := Grey; Result.A := A;
+end;
+
+function SDLFPoint(const X, Y: CFloat): TSDL_FPoint;
+begin
+  Result.X := X; Result.Y := Y;
+end;
+
+function SDLFRect(const X, Y, W, H: CFloat): TSDL_FRect;
+begin
+  Result.X := X; Result.Y := Y; Result.W := W; Result.H := H;
+end;
+
+function SDLFSegment(const X1, Y1, X2, Y2: CFloat): TCHXSDLFSegment;
+begin
+  Result.P1.X := X1; Result.P1.Y := Y1;
+  Result.P2.X := X2; Result.P2.Y := Y2;
+end;
+
+function SDLFSegment(const aP1, aP2: TSDL_FPoint) : TCHXSDLFSegment;
+begin
+  Result.P1 := aP1; Result.P2 := aP2;
 end;
 
 (*
-operator = (const a, b: TSDL_Color): Boolean;
-begin
-  // Si son totalmente transparentes les consideramos iguales
-  if (a.A = 0) and (b.A = 0) then Exit(True);
-  Result := (a.R = b.R) and (a.G = b.G) and (a.G = b.G) and (a.A = b.A);
-end;
-
-operator = (const a, b: TSDL_Point): Boolean;
-begin
-  Result := (a.X = b.X) and (a.Y = b.Y);
-end;
-
-operator = (const a, b: TSDL_FPoint): Boolean;
-begin
-  // ToDo: ¿Hacer estricto?
-  Result := SameValue(a.X, b.X) and SameValue(a.Y, b.Y);
-end;
-
-operator = (const a, b: TSDL_Rect): Boolean;
-begin
-  Result := (a.X = b.X) and (a.Y = b.Y) and (a.W = b.W) and (a.H = b.H);
-end;
-
-operator = (const a, b: TSDL_FRect): Boolean;
-begin
-  // ToDo: ¿Hacer estricto?
-  Result := SameValue(a.X, b.X) and SameValue(a.Y, b.Y) and SameValue(a.W, b.W)
-    and SameValue(a.H, b.H)
-end;
-
-operator = (const a, b: TSDL_Vertex): Boolean;
-begin
-  Result := (a.Position = b.Position) and (a.Color = a.Color) 
-    and (a.Tex_Coord = a.Tex_Coord);
-end;
-*)
-
 function SDLColor2Str(aColor : TSDL_Color) : String;
 begin
   Result := Format('%0:d, %1:d, %2:d, %3:d',
@@ -186,5 +297,5 @@ begin
   if Length(Components) < 4 then Exit;
   Result.A := EnsureRange(StrToInt(Components[3]), 0, 255);
 end;
-
+*)
 end.